Fix: Address feedback for safe_eval function
